Implemented [Issue-4] Android Service for telemetry logging (AccelerometerTelemetryService).
