+++ android-sensor-app/app/src/main/AndroidManifest.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ragres.mongodb.iotexample">
+    package="com.ragres.mongodb.iotexample" >
 
     <application
         android:name=".AndroidApplication"
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme" >
         <activity
             android:name=".ui.activities.MainActivity"
-            android:label="@string/app_name">
+            android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -19,8 +19,12 @@
         </activity>
 
         <!-- MQTT Service. -->
-        <service android:name="org.eclipse.paho.android.service.MqttService"></service>
-
+        <service android:name="org.eclipse.paho.android.service.MqttService" />
+        <service
+            android:name=".services.AccelerometerTelemetryService"
+            android:enabled="true"
+            android:exported="true" >
+        </service>
     </application>
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
