+++ android-sensor-app/app/src/main/java/com/ragres/mongodb/iotexample/AndroidApplication.java
@@ -2,6 +2,7 @@ package com.ragres.mongodb.iotexample;
 
 import android.app.Application;
 import android.content.Context;
+import android.content.Intent;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
@@ -13,6 +14,7 @@ import com.ragres.mongodb.iotexample.controllers.ConnectivityController;
 import com.ragres.mongodb.iotexample.domain.dto.SensorDataDTO;
 import com.ragres.mongodb.iotexample.domain.dto.payloads.AccelerometerDataPayload;
 import com.ragres.mongodb.iotexample.misc.Logging;
+import com.ragres.mongodb.iotexample.services.AccelerometerTelemetryService;
 
 import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
@@ -25,62 +27,8 @@ public class AndroidApplication extends Application {
     public static final String SUBTOPIC_ACCELEROMETER = "/accelerometer";
     public static final String SUBTOPIC_DEBUG = "/debug";
     public static final String SUBTOPIC_CONNECTED = "/connected";
-    /**
-     * Is logging of sensor data enabled?
-     * If true, accelerometer data is written to log.
-     */
-    private static boolean LOG_SENSOR_DATA = false;
-    /**
-     * Listener for accelerometer sensor data.
-     */
-    private SensorEventListener accelerometerListener = new SensorEventListener() {
-        @Override
-        public void onSensorChanged(SensorEvent event) {
-
-            // ASSUMPTION: we only want to handle accelerometer sensor data.
-            if (Sensor.TYPE_ACCELEROMETER != event.sensor.getType()) {
-                return;
-            }
-
-            float x = event.values[0];
-            float y = event.values[1];
-            float z = event.values[2];
-
-            if (LOG_SENSOR_DATA) {
-                Log.v(Logging.TAG, "Accelerometer data: x=" + String.valueOf(x) +
-                        ", y=" + String.valueOf(y) + ", z=" + String.valueOf(z));
-            }
-
-            // ASSUMPTION: If mqttClient is != null, it is connected
-            // (as defined by connection handling code).
-            // TODO/DISCUSSION: methods from outer class are accessed without 'this'.
-            if (isSendSensorDataEnabled() && null != getConnectivityController().getMqttClient()) {
-
-                AccelerometerDataPayload accelerometerData = AccelerometerDataPayload.fromArray(event.values);
-                SensorDataDTO sensorDataDTO = SensorDataDTO.
-                        createWithPayload(accelerometerData);
-                String jsonData = gson.toJson(sensorDataDTO);
-
-                MqttMessage mqttMessage = new MqttMessage(jsonData.getBytes());
-                try {
-                    String topic = getDeviceSubTopic(SUBTOPIC_ACCELEROMETER);
-                    getConnectivityController().getMqttClient().publish(
-                            topic,
-                            mqttMessage);
-                } catch (MqttException e) {
-                    Log.e(Logging.TAG, e.toString());
-                }
-            }
 
 
-        }
-
-
-        @Override
-        public void onAccuracyChanged(Sensor sensor, int accuracy) {
-            // Unused.
-        }
-    };
     /**
      * Device name.
      * ASSUMPTION: this is usable as MQTT identifier.
@@ -90,18 +38,7 @@ public class AndroidApplication extends Application {
      * Connectivity controller.
      */
     private ConnectivityController connectivityController;
-    /**
-     * JSON serializer.
-     */
-    private Gson gson = new Gson();
-    /**
-     * Android sensor manager.
-     */
-    private SensorManager sensorManager;
-    /**
-     * Accelerometer sensor.
-     */
-    private Sensor accelerometerSensor;
+
     /**
      * Is transmission of sensor data over MQTT
      * enabled?
@@ -129,10 +66,8 @@ public class AndroidApplication extends Application {
 
         this.connectivityController = new ConnectivityController(this);
 
-        this.sensorManager = (SensorManager) getSystemService(Context.SENSOR_SERVICE);
-        this.accelerometerSensor = this.sensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
-        this.sensorManager.registerListener(this.accelerometerListener, this.accelerometerSensor,
-                SensorManager.SENSOR_DELAY_NORMAL);
+        Intent startServiceIntent = new Intent(this, AccelerometerTelemetryService.class);
+        this.startService(startServiceIntent);
     }
 
     /**
